refactor: devided to components
