+++ src/components/orderPage/Result/style.scss
@@ -1,0 +1,82 @@
+@import '../../../styles/variables.scss';
+
+.result {
+  width: 33%;
+  padding-left: 2%;
+  padding-top: 33px;
+  font-weight: 300;
+  font-size: 14px;
+  line-height: 16px;
+  color: $black;
+  border-left: solid 1px $gray;
+
+  & .list {
+    margin-bottom: 10px;
+
+    & .dots { 
+      background: url('https://i.stack.imgur.com/otJN0.png') repeat-x bottom; 
+    }
+
+    & .field {
+      background-color: #FFFFFF;
+    }
+
+    & .link {
+      width: 150px;
+      display: inline-block;
+    }
+
+    & .address {
+      color: $gray;
+    }
+  }
+
+  & .price {
+    margin-top: 32px;
+    color: $black;
+    font-weight: 500;
+
+    & span {
+      font-weight: 700;
+    }
+  }
+
+  & .btn {
+    border-radius: 8px;
+    width: 287px;
+    height: 48px;
+    color: $white;
+    font-weight: 500;
+    font-size: 18px;
+    line-height: 21px;
+    text-align: center;
+    margin-top: 32px;
+  }
+
+  & .btn-disable {
+    background: $grayFooter;
+  }
+
+  & .btn-active {
+    background: $mainAccent;
+  }
+}
+
+@media (max-width: 1050px) {
+  .result {
+    width: 100%;
+    margin-left: 60px;
+    margin-bottom: 30px;
+    margin-top: 30px;
+    padding: 0px;
+    border-left: none;
+  }
+}
+
+@media (max-width: 500px) {
+  .result {
+    width: 100%;
+    margin-left: 10px;
+    margin-right: 10px;
+  }
+}
