+++ src/components/orderPage/Tabs/style.scss
@@ -1,0 +1,85 @@
+@import '../../../styles/variables.scss';
+
+.statuses {
+  display: flex;
+  align-items: center;
+  width: 100%;
+  background: $white;
+  border-top: 1px solid $grayFooter;
+  border-bottom: 1px solid $grayFooter;
+  max-height: 40px;
+
+  & .status {
+    color: $gray;
+    padding: 8px 38px 7px 0px;
+    font-weight: bold;
+    font-size: 14px;
+    line-height: 16px;
+    position: relative;
+    cursor: pointer;
+
+    &:hover {
+      opacity: 0.7;
+    }
+
+    &:first-of-type {
+      padding-left: 48px;
+    }
+
+    &::after {
+      content: ' ';
+      display: inline-block;
+      position: absolute;
+      width: 6px;
+      height: 8px;
+      right: 16px;
+      bottom: 12px;
+      background-image: url('../../../assets/orderPage/arrow.svg');
+      background-repeat: no-repeat;
+    }
+  }
+
+  & .status-active {
+    color: $mainAccent;
+  }
+
+  & .status-prev {
+    color: $black;
+  }
+}
+
+@media (max-width: 1050px) {
+  .statuses {
+    & .status {
+      justify-content: space-between;
+      align-items: center;
+
+      &::after {
+        background-image: none;
+      }
+    }
+  }
+}
+
+@media (max-width: 500px) {
+  .statuses {
+    font-weight: 400;
+    font-size: 12px;
+
+    & .status {
+      padding-right: 10px;
+      
+      &:first-of-type {
+        padding-left: 10px;
+      }
+    }
+
+    & .status-mobile {
+      display: none;
+    }
+
+    & .status-active {
+      display: block;
+    }
+  }
+}
