+++ src/components/orderPage/style.scss
@@ -13,54 +13,6 @@
 
     & .ant-layout-content {
       flex: none;
-
-      & .statuses {
-        display: flex;
-        align-items: center;
-        width: 100%;
-        background: $white;
-        border-top: 1px solid $grayFooter;
-        border-bottom: 1px solid $grayFooter;
-        max-height: 40px;
-  
-        & .status {
-          color: $gray;
-          padding: 8px 38px 7px 0px;
-          font-weight: bold;
-          font-size: 14px;
-          line-height: 16px;
-          position: relative;
-          cursor: pointer;
-  
-          &:hover {
-            opacity: 0.7;
-          }
-  
-          &:first-of-type {
-            padding-left: 48px;
-          }
-  
-          &::after {
-            content: ' ';
-            display: inline-block;
-            position: absolute;
-            width: 6px;
-            height: 8px;
-            right: 16px;
-            bottom: 12px;
-            background-image: url('../../assets/orderPage/arrow.svg');
-            background-repeat: no-repeat;
-          }
-        }
-  
-        & .status-active {
-          color: $mainAccent;
-        }
-  
-        & .status-prev {
-          color: $black;
-        }
-      }
     }
 
     & .content {
@@ -75,68 +27,6 @@
         width: 65%;
         margin-top: 20px;
       }
-  
-      & .result {
-        width: 33%;
-        padding-left: 2%;
-        padding-top: 33px;
-        font-weight: 300;
-        font-size: 14px;
-        line-height: 16px;
-        color: $black;
-        border-left: solid 1px $gray;
-
-        & .list {
-          margin-bottom: 10px;
-
-          & .dots { 
-            background: url('https://i.stack.imgur.com/otJN0.png') repeat-x bottom; 
-          }
-    
-          & .field {
-            background-color: #FFFFFF;
-          }
-    
-          & .link {
-            width: 150px;
-            display: inline-block;
-          }
-    
-          & .address {
-            color: $gray;
-          }
-        }
-  
-        & .price {
-          margin-top: 32px;
-          color: $black;
-          font-weight: 500;
-  
-          & span {
-            font-weight: 700;
-          }
-        }
-  
-        & .btn {
-          border-radius: 8px;
-          width: 287px;
-          height: 48px;
-          color: $white;
-          font-weight: 500;
-          font-size: 18px;
-          line-height: 21px;
-          text-align: center;
-          margin-top: 32px;
-        }
-
-        & .btn-disable {
-          background: $grayFooter;
-        }
-
-        & .btn-active {
-          background: $mainAccent;
-        }
-      }
     }
   }
 }
@@ -144,29 +34,9 @@
 @media (max-width: 1050px) {
   .order-page {
     & .wrapper {
-      & .ant-layout {
-        & .statuses {
-          & .status {
-            justify-content: space-between;
-            align-items: center;
-    
-            &::after {
-              background-image: none;
-            }
-          }
-        }
-    
+      & .ant-layout {    
         & .forms {
           width: 100%;
-        }
-    
-        & .result {
-          width: 100%;
-          margin-left: 60px;
-          margin-bottom: 30px;
-          margin-top: 30px;
-          padding: 0px;
-          border-left: none;
         }
       }
     }
@@ -179,35 +49,6 @@
 
     & .wrapper {
       & .ant-layout {
-        & .tabs {
-
-          & .statuses {
-            font-weight: 400;
-            font-size: 12px;
-
-            & .status {
-              padding-right: 10px;
-              
-              &:first-of-type {
-                padding-left: 10px;
-              }
-            }
-
-            & .status-mobile {
-              display: none;
-            }
-      
-            & .status-active {
-              display: block;
-            }
-          }
-  
-          & .result {
-            width: 100%;
-            margin-left: 10px;
-            margin-right: 10px;
-          }
-        }
 
         & .content {
           z-index: 10;
